+++ src/app/Http/Controllers/CrudController.php
@@ -8,13 +8,11 @@ use Illuminate\Http\Request as StoreRequest;
 use Illuminate\Support\Facades\Form as Form;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Http\Request as UpdateRequest;
-// CRUD Traits for non-core features
 use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Backpack\CRUD\app\Http\Controllers\CrudFeatures\Reorder;
 use Backpack\CRUD\app\Http\Controllers\CrudFeatures\AjaxTable;
 use Backpack\CRUD\app\Http\Controllers\CrudFeatures\Revisions;
-// Default request classes
 use Backpack\CRUD\app\Http\Controllers\CrudFeatures\SaveActions;
 use Backpack\CRUD\app\Http\Controllers\CrudFeatures\ShowDetailsRow;